+++ src/Settings.ts
@@ -19,11 +19,11 @@ export class Settings {
   static get pngIconColor() {
     return Settings.getSettings('pngIconColor') as string;
   }
-  static get customIconsArchivePath() {
-    return Settings.getSettings('customIconsArchivePath') as string;
+  static get customIconsArchivePaths() {
+    return Settings.getSettings('customIconsArchivePaths') as string[];
   }
-  static get customIconsFolderPath() {
-    return Settings.getSettings('customIconsFolderPath') as string;
+  static get customIconsFolderPaths() {
+    return Settings.getSettings('customIconsFolderPaths') as string[];
   }
   static get customIcons() {
     return Settings.getSettings('customIcons') as string;